toolkit/components/extensions/test/mochitest/test_ext_schema.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for schema API creation</title>
  <script type="text/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="chrome://mochikit/content/tests/SimpleTest/SpawnTask.js"></script>
  <script type="text/javascript" src="chrome://mochikit/content/tests/SimpleTest/ExtensionTestUtils.js"></script>
  <script type="text/javascript" src="chrome_head.js"></script>
  <script type="text/javascript" src="head.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
</head>
<body>

<script type="text/javascript">
"use strict";

add_task(async function testEmptySchema() {
  function background() {
    browser.test.assertEq(undefined, browser.manifest, "browser.manifest is not defined");
    browser.test.assertTrue(!!browser.storage, "browser.storage should be defined");
    browser.test.assertEq(undefined, browser.contextMenus, "browser.contextMenus should not be defined");
    browser.test.notifyPass("schema");
  }

  let extension = ExtensionTestUtils.loadExtension({
    background,
    manifest: {
      permissions: ["storage"],
    },
  });

  await extension.startup();
  await extension.awaitFinish("schema");
  await extension.unload();
});

add_task(async function testUnknownProperties() {
  function background() {
    browser.test.notifyPass("loaded");
  }

  let extension = ExtensionTestUtils.loadExtension({
    manifest: {
      permissions: ["unknownPermission"],

      unknown_property: {},
    },

    background,
  });

  let messages = [
    {message: /processing permissions\.0: Value "unknownPermission"/},
    {message: /processing unknown_property: An unexpected property was found in the WebExtension manifest/},
  ];

  let waitForConsole = new Promise(resolve => {
    SimpleTest.monitorConsole(resolve, messages);
  });

  await extension.startup();

  await extension.awaitFinish("loaded");

  await extension.unload();

  SimpleTest.endMonitorConsole();
  await waitForConsole;
});
</script>

</body>
</html>
